# Optimize reading "coordinates" arrays.