… src/NetTopologySuite.IO.GeoJSON4STJ/Converters/StjGeometryConverter.Coordinates.cs
… src/NetTopologySuite.IO.GeoJSON4STJ/Converters/Utility.cs
… src/NetTopologySuite.IO.GeoJSON4STJ/Properties/Resources.Designer.cs