Following up on https://github.com/NetTopologySuite/NetTopologySuite.IO.GeoJSON/issues/44#issuecomment-584110632