@airbreather 2020-02-15
I downloaded all the files from https://github.com/Azure-Samples/AzureMapsCodeSamples/tree/master/AzureMapsCodeSamples/Common/data/geojson (and its subfolder) and ran a benchmark to compare.  The benchmark source code was the same for both and can be found [here](https://gist.github.com/airbreather/581123e2f5cfe62f337bf60cc8537131).

I also had to write a converter for `FeatureCollection`, at least enough to get the rest of this running.  The implementation isn't **particularly** interesting for the purposes of this PR, but it does show [how](https://gist.github.com/airbreather/0a960ba956f6bde69b548ed027304979#file-stjfeaturecollectionconverter-cs-L38) to read nested objects using the converters that are already in `options` instead of using special instances of our converters directly... source code for that is [here](https://gist.github.com/airbreather/0a960ba956f6bde69b548ed027304979).

Anyway here are the results:

# Before
(with the changes to `Utility.cs` and `Resources` so the `FeatureCollection` converter can work)

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i7-6850K CPU 3.60GHz (Skylake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  Job-NMJNND : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT

Server=True  

```
| Method |                 file |           Mean |        Error |       StdDev |         Median |    Gen 0 |  Gen 1 | Gen 2 |    Allocated |
|------- |--------------------- |---------------:|-------------:|-------------:|---------------:|---------:|-------:|------:|-------------:|
| **RunSTJ** |             **GpsTrace** |     **6,076.4 us** |     **35.71 us** |     **33.40 us** |     **6,071.1 us** |  **15.6250** | **7.8125** |     **-** |   **1691.08 KB** |
| **RunSTJ** | **Ipswich Road Network** |   **350,506.9 us** |  **1,267.16 us** |  **1,185.30 us** |   **350,206.6 us** |        **-** |      **-** |     **-** |  **37884.09 KB** |
| **RunSTJ** |     **SamplePoiDataSet** |    **18,761.3 us** |     **87.21 us** |     **81.58 us** |    **18,757.5 us** |  **31.2500** |      **-** |     **-** |   **5779.22 KB** |
| **RunSTJ** | **USGS_(...)tours [25]** |       **682.1 us** |      **1.62 us** |      **1.51 us** |       **681.6 us** |   **0.9766** |      **-** |     **-** |     **94.19 KB** |
| **RunSTJ** | **US_County_Boundaries** |   **104,173.5 us** |    **219.82 us** |    **194.87 us** |   **104,117.2 us** |        **-** |      **-** |     **-** |  **15653.55 KB** |
| **RunSTJ** | **US_Co(...)_2017 [27]** |    **62,965.1 us** |    **230.79 us** |    **180.19 us** |    **62,964.6 us** | **125.0000** |      **-** |     **-** |  **11762.75 KB** |
| **RunSTJ** |       **US_States_500k** |    **79,054.4 us** |    **228.55 us** |    **213.79 us** |    **79,033.7 us** |        **-** |      **-** |     **-** |   **9202.56 KB** |
| **RunSTJ** | **US_St(...)nsity [28]** |     **3,265.8 us** |      **2.97 us** |      **2.78 us** |     **3,266.2 us** |   **3.9063** |      **-** |     **-** |    **433.83 KB** |
| **RunSTJ** |              **address** |   **992,693.2 us** | **19,148.88 us** | **24,898.95 us** | **1,006,014.9 us** |        **-** |      **-** |     **-** | **332091.03 KB** |
| **RunSTJ** | **marit(...)outes [21]** |    **64,775.4 us** |  **1,279.73 us** |  **1,992.39 us** |    **65,774.5 us** | **125.0000** |      **-** |     **-** |  **15040.97 KB** |
| **RunSTJ** |       **montreal_roads** |   **800,690.2 us** | **15,340.63 us** | **20,479.30 us** |   **808,330.9 us** |        **-** |      **-** |     **-** | **141791.27 KB** |
| **RunSTJ** |         **niagra_roads** |   **483,457.6 us** |  **9,309.79 us** | **13,351.82 us** |   **487,800.6 us** |        **-** |      **-** |     **-** |  **76203.66 KB** |
| **RunSTJ** |              **parcels** | **3,689,803.3 us** |  **2,845.35 us** |  **2,661.54 us** | **3,689,474.8 us** |        **-** |      **-** |     **-** | **301628.86 KB** |
| **RunSTJ** |              **streets** |   **610,365.6 us** |  **2,095.14 us** |  **1,959.80 us** |   **609,730.4 us** |        **-** |      **-** |     **-** |  **72562.89 KB** |
| **RunSTJ** | **trans(...)atasd [21]** |   **127,714.0 us** |    **249.88 us** |    **221.52 us** |   **127,744.8 us** |        **-** |      **-** |     **-** |  **14402.99 KB** |

# After
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i7-6850K CPU 3.60GHz (Skylake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  Job-FSNEJR : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT

Server=True  

```
| Method |                 file |           Mean |        Error |       StdDev |         Median |    Gen 0 |  Gen 1 | Gen 2 |    Allocated |
|------- |--------------------- |---------------:|-------------:|-------------:|---------------:|---------:|-------:|------:|-------------:|
| **RunSTJ** |             **GpsTrace** |     **5,061.8 us** |     **18.80 us** |     **16.66 us** |     **5,058.7 us** |  **15.6250** | **7.8125** |     **-** |   **1537.34 KB** |
| **RunSTJ** | **Ipswich Road Network** |   **294,260.1 us** |  **5,854.08 us** |  **7,611.96 us** |   **293,035.7 us** |        **-** |      **-** |     **-** |  **30507.09 KB** |
| **RunSTJ** |     **SamplePoiDataSet** |    **16,755.5 us** |     **63.64 us** |     **56.41 us** |    **16,744.8 us** |  **31.2500** |      **-** |     **-** |   **5475.68 KB** |
| **RunSTJ** | **USGS_(...)tours [25]** |       **398.4 us** |      **0.94 us** |      **0.84 us** |       **398.5 us** |   **0.4883** |      **-** |     **-** |     **68.71 KB** |
| **RunSTJ** | **US_County_Boundaries** |    **66,380.8 us** |    **356.07 us** |    **333.07 us** |    **66,273.9 us** | **125.0000** |      **-** |     **-** |  **13506.75 KB** |
| **RunSTJ** | **US_Co(...)_2017 [27]** |    **43,499.5 us** |    **190.55 us** |    **178.24 us** |    **43,524.7 us** |  **83.3333** |      **-** |     **-** |  **10760.53 KB** |
| **RunSTJ** |       **US_States_500k** |    **45,042.8 us** |     **99.26 us** |     **82.89 us** |    **45,008.9 us** |        **-** |      **-** |     **-** |   **7008.36 KB** |
| **RunSTJ** | **US_St(...)nsity [28]** |     **1,975.1 us** |      **4.62 us** |      **4.32 us** |     **1,975.0 us** |        **-** |      **-** |     **-** |    **365.55 KB** |
| **RunSTJ** |              **address** |   **930,273.6 us** | **17,393.72 us** | **16,270.10 us** |   **938,223.4 us** |        **-** |      **-** |     **-** | **321946.67 KB** |
| **RunSTJ** | **marit(...)outes [21]** |    **52,825.7 us** |  **1,046.38 us** |  **1,776.83 us** |    **51,319.9 us** | **111.1111** |      **-** |     **-** |  **14122.61 KB** |
| **RunSTJ** |       **montreal_roads** |   **703,333.5 us** | **13,303.56 us** | **14,234.67 us** |   **709,995.6 us** |        **-** |      **-** |     **-** |  **130679.2 KB** |
| **RunSTJ** |         **niagra_roads** |   **408,388.0 us** |  **7,844.18 us** |  **9,920.38 us** |   **412,823.6 us** |        **-** |      **-** |     **-** |  **68810.37 KB** |
| **RunSTJ** |              **parcels** | **3,027,714.3 us** |  **3,189.51 us** |  **2,663.38 us** | **3,027,032.9 us** |        **-** |      **-** |     **-** | **235603.76 KB** |
| **RunSTJ** |              **streets** |   **511,798.1 us** |  **1,851.40 us** |  **1,641.22 us** |   **511,465.8 us** |        **-** |      **-** |     **-** |  **63116.63 KB** |
| **RunSTJ** | **trans(...)atasd [21]** |    **81,521.1 us** |    **359.23 us** |    **336.02 us** |    **81,526.5 us** |        **-** |      **-** |     **-** |  **12049.58 KB** |

# One table
manually weaved the rows in case it makes this easier to read

| Method |                 file |           Mean |        Error |       StdDev |         Median |    Gen 0 |  Gen 1 | Gen 2 |    Allocated |
|------- |--------------------- |---------------:|-------------:|-------------:|---------------:|---------:|-------:|------:|-------------:|
| before |             GpsTrace |     6,076.4 us |     35.71 us |     33.40 us |     6,071.1 us |  15.6250 | 7.8125 |     - |   1691.08 KB |
| after  |             GpsTrace |     5,061.8 us |     18.80 us |     16.66 us |     5,058.7 us |  15.6250 | 7.8125 |     - |   1537.34 KB |
| before | Ipswich Road Network |   350,506.9 us |  1,267.16 us |  1,185.30 us |   350,206.6 us |        - |      - |     - |  37884.09 KB |
| after  | Ipswich Road Network |   294,260.1 us |  5,854.08 us |  7,611.96 us |   293,035.7 us |        - |      - |     - |  30507.09 KB |
| before |     SamplePoiDataSet |    18,761.3 us |     87.21 us |     81.58 us |    18,757.5 us |  31.2500 |      - |     - |   5779.22 KB |
| after  |     SamplePoiDataSet |    16,755.5 us |     63.64 us |     56.41 us |    16,744.8 us |  31.2500 |      - |     - |   5475.68 KB |
| before | USGS_(...)tours [25] |       682.1 us |      1.62 us |      1.51 us |       681.6 us |   0.9766 |      - |     - |     94.19 KB |
| after  | USGS_(...)tours [25] |       398.4 us |      0.94 us |      0.84 us |       398.5 us |   0.4883 |      - |     - |     68.71 KB |
| before | US_County_Boundaries |   104,173.5 us |    219.82 us |    194.87 us |   104,117.2 us |        - |      - |     - |  15653.55 KB |
| after  | US_County_Boundaries |    66,380.8 us |    356.07 us |    333.07 us |    66,273.9 us | 125.0000 |      - |     - |  13506.75 KB |
| before | US_Co(...)_2017 [27] |    62,965.1 us |    230.79 us |    180.19 us |    62,964.6 us | 125.0000 |      - |     - |  11762.75 KB |
| after  | US_Co(...)_2017 [27] |    43,499.5 us |    190.55 us |    178.24 us |    43,524.7 us |  83.3333 |      - |     - |  10760.53 KB |
| before |       US_States_500k |    79,054.4 us |    228.55 us |    213.79 us |    79,033.7 us |        - |      - |     - |   9202.56 KB |
| after  |       US_States_500k |    45,042.8 us |     99.26 us |     82.89 us |    45,008.9 us |        - |      - |     - |   7008.36 KB |
| before | US_St(...)nsity [28] |     3,265.8 us |      2.97 us |      2.78 us |     3,266.2 us |   3.9063 |      - |     - |    433.83 KB |
| after  | US_St(...)nsity [28] |     1,975.1 us |      4.62 us |      4.32 us |     1,975.0 us |        - |      - |     - |    365.55 KB |
| before |              address |   992,693.2 us | 19,148.88 us | 24,898.95 us | 1,006,014.9 us |        - |      - |     - | 332091.03 KB |
| after  |              address |   930,273.6 us | 17,393.72 us | 16,270.10 us |   938,223.4 us |        - |      - |     - | 321946.67 KB |
| before | marit(...)outes [21] |    64,775.4 us |  1,279.73 us |  1,992.39 us |    65,774.5 us | 125.0000 |      - |     - |  15040.97 KB |
| after  | marit(...)outes [21] |    52,825.7 us |  1,046.38 us |  1,776.83 us |    51,319.9 us | 111.1111 |      - |     - |  14122.61 KB |
| before |       montreal_roads |   800,690.2 us | 15,340.63 us | 20,479.30 us |   808,330.9 us |        - |      - |     - | 141791.27 KB |
| after  |       montreal_roads |   703,333.5 us | 13,303.56 us | 14,234.67 us |   709,995.6 us |        - |      - |     - |  130679.2 KB |
| before |         niagra_roads |   483,457.6 us |  9,309.79 us | 13,351.82 us |   487,800.6 us |        - |      - |     - |  76203.66 KB |
| after  |         niagra_roads |   408,388.0 us |  7,844.18 us |  9,920.38 us |   412,823.6 us |        - |      - |     - |  68810.37 KB |
| before |              parcels | 3,689,803.3 us |  2,845.35 us |  2,661.54 us | 3,689,474.8 us |        - |      - |     - | 301628.86 KB |
| after  |              parcels | 3,027,714.3 us |  3,189.51 us |  2,663.38 us | 3,027,032.9 us |        - |      - |     - | 235603.76 KB |
| before |              streets |   610,365.6 us |  2,095.14 us |  1,959.80 us |   609,730.4 us |        - |      - |     - |  72562.89 KB |
| after  |              streets |   511,798.1 us |  1,851.40 us |  1,641.22 us |   511,465.8 us |        - |      - |     - |  63116.63 KB |
| before | trans(...)atasd [21] |   127,714.0 us |    249.88 us |    221.52 us |   127,744.8 us |        - |      - |     - |  14402.99 KB |
| after  | trans(...)atasd [21] |    81,521.1 us |    359.23 us |    336.02 us |    81,526.5 us |        - |      - |     - |  12049.58 KB |