@airbreather 2020-02-17
FWIW, I extracted all the geometries from the features and collected them all into GeoJSON files each containing a single `MultiX` geometry to get a proper comparison of just this change... results:

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i7-6850K CPU 3.60GHz (Skylake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  Job-VEVBPV : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT

Server=True  

```

| Method |                       source |           Mean |       Error |      StdDev |   Gen 0 |   Gen 1 |   Gen 2 |    Allocated |
|------- |----------------------------- |---------------:|------------:|------------:|--------:|--------:|--------:|-------------:|
| before |                     GpsTrace |     1,401.6 us |    44.37 us |    51.09 us |  3.9063 |  1.9531 |       - |     524.8 KB |
|  after |                     GpsTrace |       697.1 us |     6.46 us |     6.04 us |  4.8828 |  1.9531 |       - |    492.76 KB |
| before |         Ipswich Road Network |   242,646.9 us | 4,210.63 us | 3,516.06 us |       - |       - |       - |  20403.49 KB |
|  after |         Ipswich Road Network |   186,107.6 us | 1,580.44 us | 1,478.35 us |       - |       - |       - |   6932.14 KB |
| before |             SamplePoiDataSet |     2,659.7 us |    19.77 us |    17.53 us | 11.7188 |  7.8125 |  3.9063 |    1033.3 KB |
|  after |             SamplePoiDataSet |     1,347.0 us |     8.08 us |     7.16 us |  7.8125 |  3.9063 |       - |    969.22 KB |
| before |    USGS_M7EarthquakeContours |       549.1 us |     0.56 us |     0.49 us |       - |       - |       - |     86.67 KB |
|  after |    USGS_M7EarthquakeContours |       259.4 us |     1.82 us |     1.52 us |       - |       - |       - |     43.38 KB |
| before |         US_County_Boundaries |    68,865.9 us |   154.05 us |   144.09 us |       - |       - |       - |   12268.6 KB |
|  after |         US_County_Boundaries |    32,693.2 us |    60.44 us |    53.58 us |       - |       - |       - |   4793.58 KB |
| before |  US_County_Unemployment_2017 |    39,378.7 us |    73.59 us |    68.83 us | 76.9231 | 76.9231 | 76.9231 |   7538.87 KB |
|  after |  US_County_Unemployment_2017 |    19,141.4 us |   183.97 us |   172.09 us |       - |       - |       - |   3010.92 KB |
| before |               US_States_500k |    60,229.8 us |   156.07 us |   145.99 us |       - |       - |       - |   9845.79 KB |
|  after |               US_States_500k |    29,096.8 us |    41.59 us |    36.87 us | 31.2500 |       - |       - |   3949.57 KB |
| before | US_States_Population_Density |     2,517.8 us |     5.19 us |     4.60 us |  3.9063 |  3.9063 |  3.9063 |    404.08 KB |
|  after | US_States_Population_Density |     1,167.9 us |     4.07 us |     3.39 us |  1.9531 |       - |       - |       200 KB |
| before |                      address |   169,771.1 us | 2,000.45 us | 1,670.47 us |       - |       - |       - |  29812.99 KB |
|  after |                      address |   133,254.5 us |   683.98 us |   639.79 us |       - |       - |       - |  25088.46 KB |
| before |        maritime-trade-routes |    19,915.1 us |    72.98 us |    68.27 us | 62.5000 | 31.2500 | 31.2500 |   4586.66 KB |
|  after |        maritime-trade-routes |     9,458.4 us |   178.73 us |   198.66 us |       - |       - |       - |   2456.77 KB |
| before |               montreal_roads |   332,108.7 us |   848.02 us |   793.23 us |       - |       - |       - |   35180.6 KB |
|  after |               montreal_roads |   250,967.2 us | 1,191.12 us | 1,055.90 us |       - |       - |       - |  14684.62 KB |
| before |                 niagra_roads |   239,689.4 us |   424.11 us |   354.15 us |       - |       - |       - |  23658.63 KB |
|  after |                 niagra_roads |   194,334.3 us | 2,607.71 us | 2,439.25 us |       - |       - |       - |      9034 KB |
| before |                      parcels | 2,848,797.4 us | 1,855.13 us | 1,735.29 us |       - |       - |       - | 244180.57 KB |
|  after |                      parcels | 2,218,798.8 us | 6,389.69 us | 5,976.92 us |       - |       - |       - |  77057.42 KB |
| before |                      streets |   399,608.6 us | 1,642.25 us | 1,536.16 us |       - |       - |       - |  32436.62 KB |
|  after |                      streets |   316,841.0 us |   801.71 us |   710.70 us |       - |       - |       - |   10198.2 KB |
| before |        transit_routes_datasd |   101,301.1 us |   279.59 us |   261.53 us |       - |       - |       - |   15137.6 KB |
|  after |        transit_routes_datasd |    58,015.5 us |   132.33 us |   117.30 us |       - |       - |       - |    5051.3 KB |